refactor: 가계부 예산 설정 디렉토리 구조 변경
