+++ src/client/modules/Account/AccountBudgetSetting/hooks/useBudgeSetForm.ts
@@ -1,0 +1,28 @@
+import { z } from 'zod';
+import { useForm } from 'react-hook-form';
+import { zodResolver } from '@hookform/resolvers/zod';
+
+import { typedObjectEntries } from '@/src/client/utils/type';
+import { EXPENSE_CATEGORY_MAP } from '@/src/app/ai/money-tracker/_ constants/category';
+
+const budgetSchemaShape = Object.fromEntries(
+  typedObjectEntries(EXPENSE_CATEGORY_MAP).map(([key]) => [key, z.number()]),
+);
+
+const BudgetSchema = z.object(budgetSchemaShape);
+
+export type BudgetFormType = z.infer<typeof BudgetSchema>;
+
+export const useBudgeSetForm = ({
+  defaultValues,
+}: {
+  defaultValues?: BudgetFormType;
+}) =>
+  useForm<BudgetFormType>({
+    resolver: zodResolver(BudgetSchema),
+    defaultValues: {
+      ...defaultValues,
+    },
+  });
+
+export default useBudgeSetForm;
