+++ src/client/modules/Account/AccountBudgetSetting/index.ts
@@ -1,0 +1,1 @@
+export { default } from './AccountBudgetSetting';
